docs: Update repository structure
In order to record changes in thoughts and manage source code according
to status, I update repository structure. The role of each folder is as
follows.

not-solve : Unsolved problems
solve : Solved problems, but not reviewed
review : Solved problem and reviewed it once
others with site names : Solved problem and reviewed it twice

Resolves: #1